+++ Changelog.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.4.2]
+
+### Added
+- Implemented the function getCorporationDetails
+- Implemented support for the parameters for the function getAllCommitteesAvailable
+- Implemented the getOrganizationList
+
+### Changed
+- Updated documentation for functions: getAgendasByCommittee, getAgendasByOrganization, getAgenda, getPDFDocumentUrl, mapJSONtoCommittee
+
 ## [v0.4.1]
 
 ### Fixed